design(write-top): change to use AppBar
CompleteButton 관련 로직을 useEditorCompleteButton으로 추출
