+++ app/(signed-in-only)/write/components/top.tsx
@@ -1,15 +1,19 @@
-import CompleteButton from './completeButton';
-import ModeSelector from './modeSelector';
+import ModeSelector from '@/app/(signed-in-only)/write/components/modeSelector';
+import { useEditorCompleteButton } from '@/app/(signed-in-only)/write/hooks/use-editor-complete-button';
+import { AppBar, AppBarBack } from '@/components/ui-unstable/app-bar';
+import { Button } from '@/components/ui/button';
 
 const Top = () => {
+  const { handleCompleteButtonClick } = useEditorCompleteButton();
+
   return (
-    <div className="fixed left-[50%] top-0 z-50 flex h-[5%] w-[100vw] max-w-lg -translate-x-[50%] items-center justify-between">
-      <button className="rounded bg-blue-500 px-4 py-2 text-white">
-        뒤로가기
-      </button>
+    <AppBar className="justify-between border-none bg-transparent">
+      <AppBarBack />
       <ModeSelector />
-      <CompleteButton />
-    </div>
+      <Button variant="ghost" onClick={handleCompleteButtonClick}>
+        완료
+      </Button>
+    </AppBar>
   );
 };
